User Model

+++ frontEnd/leaguelizer/src/Pages/ClubPage/CustomForm.js
@@ -1,8 +1,9 @@
-import { React, useEffect, useState, useCallback } from "react";
+import { React, useEffect, useState, useCallback, useContext } from "react";
 import { Button, Container, Grid, TextField } from "@mui/material";
 import URL_BASE from "./constants";
 import CustomTable from "../../Layouts/PageLayout/Table/CustomTable";
 import ToasterError from "../../Layouts/ErrorLayout/ToasterError";
+import authContext from "../../Context/Context";
 
 const initialMatchValue = {
     "club1": {
@@ -27,6 +28,7 @@ const initialMatchValue = {
 }
 
 export default function CustomForm(props) {
+    let {user} = useContext(authContext);
     const [clubNameValue, setClubNameValue] = useState(props.value.name);
     const [clubAnnualBudgetValue, setClubAnnualBudgetValue] = useState(props.value.annualBudget);
     const [clubStaffValue, setClubStaffValue] = useState(props.value.numberOfStadd);
@@ -56,15 +58,83 @@ export default function CustomForm(props) {
     const [ orderDirection, setOrderDirection ] = useState("asc");
     const [ pageNumber, setPageNumber ] = useState(1);
     const [ pageMax, setPageMax ] = useState(1);
-    const [ paginationValue, setPaginationValue ] = useState(12);
+    const [ paginationValue, setPaginationValue ] = useState(localStorage.getItem('paginationValue') ? JSON.parse(localStorage.getItem('paginationValue')) : 12);
 
     function validateMatch(){
         if (!/^[0-9]{1,2}-[0-9]{1,2}$/.test(scoreValue)){
             ToasterError("Incorrect/Impossible Score!");
             return false;
         }
-        if ((/^[0-9]+$/.test(club2Value)) ? club2Value : matchValue.club2.id === club2Value){
+        if (props.value.id === club2Value){
             ToasterError("The team cannot play with itself!");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(club2Value) && matchValue.club2.id === ""){
+            ToasterError("Invalid club2 id");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(stadiumValue) && matchValue.stadium.id === ""){
+            ToasterError("Invalid stadium id");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(compValue) && matchValue.competition.id === ""){
+            ToasterError("Invalid stadium id");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]*$/.test(roundValue)){
+            ToasterError("Invalid Round of play");
+            return false;
+        }
+        if (!/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/.test(dateValue)){
+            ToasterError("Date needs to have the following format: yyyy-mm-dd");
+            return false;
+        }
+        return true;
+    }
+
+    function validateClub() {
+        if (!/^[0-9]+$/.test(clubAnnualBudgetValue) | parseInt(clubAnnualBudgetValue) < 0){
+            ToasterError("Annual Budget must be a positive integer");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(clubStaffValue) | parseInt(clubStaffValue) < 0){
+            ToasterError("Number of staff must be a positive integer");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]+$/.test(clubNameValue)){
+            ToasterError("Club Name can only contain numbers and letters");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(clubStadiumValue)){
+            ToasterError("Club stadium incorrect!");
+            return false;
+        }
+        if (!/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/.test(clubDateValue)){
+            ToasterError("Date needs to have the following format: yyyy-mm-dd");
+            return false;
+        }
+        return true;
+    }
+
+    function validateCompetition() {
+        if (!/^[0-9]+$/.test(teamNumber) | parseInt(teamNumber) < 0){
+            ToasterError("There must be a positive number of teams");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(prizeMoney) | parseInt(prizeMoney) < 0){
+            ToasterError("There must be a positive prize");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]+$/.test(leagueName)){
+            ToasterError("Competition Name can only contain numbers and letters");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]+$/.test(compType)){
+            ToasterError("Competition type can only contain numbers and letters");
+            return false;
+        }
+        if (!/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/.test(foundedDate)){
+            ToasterError("Date needs to have the following format: yyyy-mm-dd");
             return false;
         }
         return true;
@@ -73,11 +143,16 @@ export default function CustomForm(props) {
     const getPageMax = useCallback(() => {
         if (props.value.id === undefined)
             return;
+
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
     
-        fetch(URL_BASE + String(props.value.id) + "/competitions/?pageNumber=0")
+        fetch(URL_BASE + String(props.value.id) + "/competitions/?pageNumber=" + String(paginationValue))
             .then(number => number.json())
             .then(number => setPageMax(number["pageNumber"]));
-    }, [props.value.id]);
+    }, [props.value.id, paginationValue]);
 
     useEffect(() => {
         setClubNameValue(props.value.name)
@@ -94,6 +169,11 @@ export default function CustomForm(props) {
     var getUrlForMatches = useCallback(() => {
         if (props.value.id === undefined)
             return "";
+
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return "";
+        }
 
         return URL_BASE + String(props.value.id) + "/competitions/?page=" + String(pageNumber) + "&pageNumber=" + String(paginationValue);
     }, [pageNumber, props.value.id, paginationValue])
@@ -169,6 +249,10 @@ export default function CustomForm(props) {
     }
 
     const postButtonHandler = () => {
+        if (!validateClub()){
+            return;
+        }
+
         const requestOptions = {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
@@ -178,7 +262,8 @@ export default function CustomForm(props) {
                 "numberOfStadd": clubStaffValue,
                 "foundedDate": clubDateValue,
                 "stadium": clubStadiumValue,
-                "league": clubLeagueValue
+                "league": clubLeagueValue,
+                "user":(user) ? user.user_id : null
             })
         };
 
@@ -193,6 +278,10 @@ export default function CustomForm(props) {
     }
 
     const postWithLeagueHandler = () => {
+        if (!validateClub || !validateCompetition){
+            return;
+        }
+
         const requestOptions = {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
@@ -202,27 +291,44 @@ export default function CustomForm(props) {
                 "numberOfStadd": clubStaffValue,
                 "foundedDate": clubDateValue,
                 "stadium": clubStadiumValue,
+                "user":(user) ? user.user_id : null,
                 "league": {
                     "name": leagueName,
                     "numberOfTeams": teamNumber,
                     "prizeMoney": prizeMoney,
                     "foundedDate": foundedDate,
-                    "competitionType": compType
+                    "competitionType": compType,
+                    "user":(user) ? user.user_id : null
                 }
             })
         };
 
-        fetch(URL_BASE + "league", requestOptions)
+        fetch(URL_BASE + "league/", requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.stadium !== undefined)
+                    ToasterError(message.stadium[0]);
+                else if (message.league !== undefined)
+                    ToasterError(message.league[0]);
+                else if (message.user !== undefined)
+                    ToasterError(message.user[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
                 else
                     props.refresh();
             })
     }
 
     const putButtonHandler = () => {
+        if (!validateClub()){
+            return;
+        }
+        
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
+
         const requestOptions = {
             method: 'PUT',
             headers: { 'Content-Type': 'application/json' },
@@ -241,14 +347,25 @@ export default function CustomForm(props) {
         fetch(URL, requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.stadium !== undefined)
+                    ToasterError(message.stadium[0]);
+                else if (message.league !== undefined)
+                    ToasterError(message.league[0]);
+                else if (message.user !== undefined)
+                    ToasterError(message.user[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
                 else
                     props.refresh();
             })
     }
 
     const deleteButtonHandler = () => {
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
+
         const requestOptions = {
             method: 'DELETE'
         };
@@ -256,21 +373,7 @@ export default function CustomForm(props) {
         const URL = URL_BASE + String(props.value.id)
 
         fetch(URL, requestOptions)
-            .then(message => message.json())
-            .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
-                else
-                    props.refresh();
-            })
-    }
-
-    const refresh = () => {
-        changeMatchValues(initialMatchValue);
-        setMatchList([]);
-        fetch(getUrlForMatches())
-            .then(match => match.json())
-            .then(match => setMatchList(match));
+            .then(() => props.refresh());
     }
 
     const postMatchButtonHandler = () => {
@@ -286,23 +389,39 @@ export default function CustomForm(props) {
                 "stadium": (/^[0-9]+$/.test(stadiumValue)) ? stadiumValue : matchValue.stadium.id,
                 "roundOfPlay": roundValue,
                 "score": scoreValue,
-                "date": dateValue
+                "date": dateValue,
+                "user":(user) ? user.user_id : null
             })
         };
 
         fetch(URL_BASE + String(props.value.id) + "/competitions/" , requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.club1 !== undefined)
+                    ToasterError(message.club1[0]);
+                else if (message.club2 !== undefined)
+                    ToasterError(message.club2[0]);
+                else if (message.stadium !== undefined)
+                    ToasterError(message.stadium[0]);
+                else if (message.competition !== undefined)
+                    ToasterError(message.competition[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
+                else if (message.user !== undefined)
+                    ToasterError(message.user);
                 else
-                    refresh();
+                    props.refresh();
             })
     }
 
     const putMatchButtonHandler = () => {
         if (!validateMatch())
             return;
+
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
 
         const requestOptions = {
             method: 'PUT',
@@ -323,14 +442,29 @@ export default function CustomForm(props) {
         fetch(URL, requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.club1 !== undefined)
+                    ToasterError(message.club1[0]);
+                else if (message.club2 !== undefined)
+                    ToasterError(message.club2[0]);
+                else if (message.stadium !== undefined)
+                    ToasterError(message.stadium[0]);
+                else if (message.competition !== undefined)
+                    ToasterError(message.competition[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
+                else if (message.user !== undefined)
+                    ToasterError(message.user);
                 else
-                    refresh();
+                    props.refresh();
             })
     }
 
     const deleteMatchButtonHandler = () => {
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
+
         const requestOptions = {
             method: 'DELETE'
         };
@@ -338,13 +472,7 @@ export default function CustomForm(props) {
         const URL = URL_BASE + String(matchValue.id) + "/competitions/"
 
         fetch(URL, requestOptions)
-            .then(message => message.json())
-            .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
-                else
-                    refresh();
-            })
+            .then(() => props.refresh());
     }
 
     return (
